emit additional files for bundle

+++ packaging.py
@@ -13,6 +13,10 @@ TARGET_DIR = THIS_DIR / BUNDLED_PACKAGES_DIR
 PLUGIN_DIR = THIS_DIR / "mi_companion"
 REQUIREMENTS_FILE = PLUGIN_DIR / "requirements.txt"
 
+BUNDLE_VERSION = "0.0.1"
+MIN_QGIS_VERSION = "3.38"
+BUNDLE_PROJECT_NAME = "MapsIndoors"
+
 logger = logging.getLogger(__name__)
 
 
@@ -42,9 +46,10 @@ def package_dependencies(
     clean: bool = True,
     python_version: str = "3.12",
 ) -> None:
-    if target_site_packages_dir.exists():
-        if clean:
-            shutil.rmtree(target_site_packages_dir)
+    if True:
+        if target_site_packages_dir.exists():
+            if clean:
+                shutil.rmtree(target_site_packages_dir)
 
     target_site_packages_dir.mkdir(parents=True, exist_ok=True)
 
@@ -57,28 +62,76 @@ def package_dependencies(
     os.environ["ZMQ_PREFIX"] = "bundled"
     os.environ["ZMQ_BUILD_DRAFT"] = "1"
 
-    catching_callable(
-        [
-            "pip",
-            "install",
-            "-U",
-            "-t",
-            f"{target_site_packages_dir}",
-            "-r",
-            f"{REQUIREMENTS_FILE}",
-            "--break-system-packages",
-            "--verbose",
-            # --platform manylinux2014_x86_64
-            # "--no-binary",
-            # "pyzmq",
-            # "--no-build-isolation",
-            "--only-binary",
-            ":all:",
-            "--implementation",
-            "cp",
-            "--python-version",
-            python_version,
-        ]
+    if True:
+        catching_callable(
+            [
+                "pip",
+                "install",
+                "-U",
+                "-t",
+                f"{target_site_packages_dir}",
+                "-r",
+                f"{REQUIREMENTS_FILE}",
+                "--break-system-packages",
+                "--verbose",
+                # --platform manylinux2014_x86_64
+                # "--no-binary",
+                # "pyzmq",
+                # "--no-build-isolation",
+                "--only-binary",
+                ":all:",
+                "--implementation",
+                "cp",
+                "--python-version",
+                python_version,
+            ]
+        )
+
+    emit_additional_bundle_files(python_version, target_site_packages_dir)
+
+
+def emit_additional_bundle_files(python_version, target_site_packages_dir):
+    with open(target_site_packages_dir / "metadata.txt", "w") as f:
+        f.write(
+            f"""[general]
+name={BUNDLE_PROJECT_NAME} Python {python_version} bundle
+qgisMinimumVersion={MIN_QGIS_VERSION}
+#qgisMaximumVersion=3.38
+version={BUNDLE_VERSION}
+
+author=heider
+email=chen@mapspeople.com
+
+description=A Python dependency bundle
+about=A bundle of Python dependencies for the MapsIndoors plugin
+
+tracker=https://github.com/MapsPeople
+repository=https://github.com/MapsPeople
+homepage=https://github.com/MapsPeople
+
+category=Web
+icon=icon.png
+
+hasProcessingProvider=no
+experimental=False
+
+# Tags are comma separated with spaces allowed
+tags=python, mapsindoors, companion
+      """
+        )
+
+    (target_site_packages_dir / "__init__.py").touch()
+
+    if (target_site_packages_dir / "LICENSE").exists():
+        shutil.rmtree(target_site_packages_dir / "LICENSE")
+    shutil.copy(
+        REQUIREMENTS_FILE.parent / "LICENSE", target_site_packages_dir / "LICENSE"
+    )
+
+    if (target_site_packages_dir / "icon.png").exists():
+        shutil.rmtree(target_site_packages_dir / "icon.png")
+    shutil.copy(
+        REQUIREMENTS_FILE.parent / "icon.png", target_site_packages_dir / "icon.png"
     )
